requirements.txt
elevenlabs==0.2.19
langchain==0.0.149
pandas~=2.2.1
Pillow==9.5.0
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.30.0
streamlit==1.22.0

reportlab~=4.1.0
openai==0.28.1
lxml~=5.2.1
imgkit==1.2.3